Modularized the code into multiple different files for better code clarity
