+++ .clangd
@@ -1,3 +1,3 @@
 CompileFlags:
-  Add: [-xc, -Wall, -I/usr/include/pipewire-0.3, -I/usr/include/spa-0.2, -D_REENTRANT, -lpipewire-0.3, -I/usr/include/gdk-pixbuf-2.0, -I/usr/include/glib-2.0, -I/usr/lib/x86_64-linux-gnu/glib-2.0/include, -I/usr/include/sysprof-6, -I/usr/include/libpng16, -I/usr/include/x86_64-linux-gnu, -I/usr/include/webp, -I/usr/include/libmount, -I/usr/include/blkid, -pthread, -lnotify, -lgdk_pixbuf-2.0, -lgio-2.0, -lgobject-2.0, -lglib-2.0]
+  Add: [-xc, -Wall, --include-directory=include, -I/usr/include/pipewire-0.3, -I/usr/include/spa-0.2, -D_REENTRANT, -lpipewire-0.3, -I/usr/include/gdk-pixbuf-2.0, -I/usr/include/glib-2.0, -I/usr/lib/x86_64-linux-gnu/glib-2.0/include, -I/usr/include/sysprof-6, -I/usr/include/libpng16, -I/usr/include/x86_64-linux-gnu, -I/usr/include/webp, -I/usr/include/libmount, -I/usr/include/blkid, -pthread, -lnotify, -lgdk_pixbuf-2.0, -lgio-2.0, -lgobject-2.0, -lglib-2.0]
   Compiler: clang
