+++ include/notification.h
@@ -1,0 +1,8 @@
+#ifndef VOLUME_NOTIFICATION
+#define VOLUME_NOTIFICATION
+
+#include "app.h"
+
+void notify_state(state_t* state);
+
+#endif // VOLUME_NOTIFICATION
