+++ include/volume_listener.h
@@ -1,0 +1,12 @@
+#ifndef VOLUME_LISTENER
+#define VOLUME_LISTENER
+
+#include "app.h"
+
+void start_listener(state_t* state);
+
+void init_listener(state_t* state);
+
+void cleanup_listener(state_t* state);
+
+#endif // !VOLUME_LISTENER
